moved some extension helper functions into their own dir, created meet request controller
